+++ frontend/src/components/Home/HomeMaintenance.css
@@ -38,15 +38,6 @@
     margin: auto;
 }
 
-/*.Mechanics_btn{*/
-/*    background-color: #F5F3F4;*/
-/*    !*border: 1px solid #660708;*!*/
-/*    color: #660708;*/
-/*    font-size: 16px;*/
-/*    font-weight: lighter;*/
-/*    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);*/
-/*}*/
-
 .Mechanics_title{
     background-color: rgba(255,255,255,.1);
 }
