+++ frontend/src/components/Terms/Terms.css
@@ -1,5 +1,5 @@
 .TermsScreen_hero-img{
-    background-image: url("../../../images/bg-graphics/screen-terms/panel.png");
+    background-image: url("./panel.png");
     background-repeat: no-repeat;
     background-attachment: fixed;
     height: 200px;
