--- frontend/src/style/flag.scss
@@ -1,35 +1,0 @@
-
-.flag{
-  position: absolute;
-  font-size: 16px;
-  left: 100px;
-  top:3px;
-  display: block;
-  height: 24px;
-  width: 100px;
-  line-height: 24px;
-  background: gold;
-  color: #000;
-  padding: 0 10px 0 10px;
-  z-index:1;
-
-  &:after{
-    position: absolute;
-    right:-5px;
-    top:0;
-    content: "";
-    border-right:  5px solid transparent;
-    border-bottom: 12px solid #4c4c4c;
-    border-top:    12px solid #4c4c4c;
-  }
-
-  &.red{
-    background: gold;
-    &:after{
-      border-color: gold transparent gold transparent;
-      //border-color: orange transparent orange transparent;
-    }
-  }
-
-
-}
